Revert "Remove ability to get drops from blocks without breaking them, add ability to get drops from blocks when converting them"
This reverts commit c4ed83d0a35d8658abbf8556660becdaa414c67c.

+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/registry/dragon/ability/block_effects/BlockBreakEffect.java
@@ -39,7 +39,9 @@ public record BlockBreakEffect(BlockPredicate validBlocks, LevelBasedValue proba
                 BlockEntity blockEntity = blockState.hasBlockEntity() ? level.getBlockEntity(position) : null;
                 Block.dropResources(blockState, level, position, blockEntity, dragon, tool.orElse(ItemStack.EMPTY));
             }
-            level.destroyBlock(position, dropLoot);
+            if (tool.isEmpty() || !dropLoot) {
+                level.destroyBlock(position, dropLoot);
+            }
         }
     }
